User commit

+++ src/main/java/com/example/demo/entity/StudentEntity.java
@@ -15,7 +15,7 @@ public class StudentEntity {
     private String email;
     private float cgpa;
 
-    public StudentEntity(Long id,String name,String email, double cgpa){
+    public StudentEntity(Long id,String name,String email, float cgpa){
         this.id = id;
         this.name = name;
         this.email = email;
@@ -35,7 +35,7 @@ public class StudentEntity {
         this.email = email;
     }
 
-    public void setCgpa(double cgpa) {
+    public void setCgpa(float cgpa) {
         this.cgpa = cgpa;
     }
 
@@ -46,7 +46,7 @@ public class StudentEntity {
     public String getEmail() {
         return email;
     }
-    public double getCgpa() {
+    public float getCgpa() {
         return cgpa;
     }
